refactor(warnings): sweep all project sources to zero compiler warnings
Fix ~90 warning sites across src/main.cpp, src/world.cpp, include/world.hpp,
src/llama_wrapper.cpp, src/town_consciousness.cpp, src/social_cognition.cpp,
src/nature_mind.cpp, src/village_mind.cpp under -Wconversion/-Wsign-conversion/
-Wshadow/-Wunused/-Wparentheses.

- main.cpp: fertilizer switch default, 7 season shadow renames, job-board j->jb,
  HTTP handler player_id/target_x/target_y casts, size_type indices, travel/
  farming/economy float casts, /cmd record latency+day casts
- world.cpp: terrain float casts, do-while parens, Vec2 d shadow -> dd, market
  flux casts, BuildingState initializer, weather_of_day_adapted param rename
- llama_wrapper.cpp: migrate to non-deprecated llama.cpp API
  (llama_model_load_from_file / llama_model_free / llama_init_from_model)
- other minds: unused-param/variable cleanups, size_type/float casts
- third-party httplib.h/llama.cpp warnings suppressed via scoped pragmas

Full build now emits zero warnings. Server smoke-tested (join/warp/move/till/
plant/sleep day-advance). Docs + MISTAKES (M12, M13) + CHANGELOG updated.

+++ src/town_consciousness.cpp
@@ -38,7 +38,7 @@ TownConsciousness::TownConsciousness(World& world, std::function<std::string(con
             current_adaptations_.horror = j.value("horror", json::object());
             current_adaptations_.performance = j.value("performance", json::object());
             current_adaptations_.day = j.value("day", world_.day);
-            current_adaptations_.consolidation_count = j.value("consolidation_count", 0);
+            current_adaptations_.consolidation_count = static_cast<uint32_t>(j.value("consolidation_count", 0));
         }
         current_adaptations_.day = world_.day;
     }
@@ -418,8 +418,8 @@ void TownConsciousness::load_memory() {
         memory_.discovered_secrets = j.value("discovered_secrets", json::array());
         memory_.performance_profile = j.value("performance_profile", json::object());
         memory_.narrative_state = j.value("narrative_state", json::object());
-        memory_.last_consolidation_day = j.value("last_consolidation_day", 0);
-        memory_.consolidation_count = j.value("consolidation_count", 0);
+        memory_.last_consolidation_day = static_cast<uint32_t>(j.value("last_consolidation_day", 0));
+        memory_.consolidation_count = static_cast<uint32_t>(j.value("consolidation_count", 0));
     }
 }
